src/requirements.txt
requests==2.32.3
PyYAML==6.0.2
logfmter==0.0.8
python-json-logger==3.2.1